* elpa-packages (emacsql): Include all connector libraries.
* elpa-packages (emacsql-mysql): Remove package.
* elpa-packages (emacsql-psql): Remove package.
* elpa-packages (emacsql-sqlite): Remove package.
* elpa-packages (emacsql-sqlite-builtin): Remove package.

This was requested in
https://lists.gnu.org/archive/html/emacs-devel/2022-12/msg00525.html.

None of these packages have been released on NonGNU ELPA yet but they
are already available from NonGNU-devel ELPA.  A single package in GNU
ELPA and GNU-devel ELPA depends on `emacsql': `triples', but it only
depends on `emacsql', not `emacsql-sqlite'.

On MELPA some packages do depend on `emacsql-sqlite', so there I was
forced to continue to distribute the connector libraries as separate
packages for now, but `emacsql' from MELPA now contains all libraries
too.

This will be fixed once the next version of `emacsql' is released,
which should happen next month.  That will also require new releases
of all the packages that currently depend on `emacsql-sqlite'.  You
can find more details about the roadmap and the complications in
https://github.com/melpa/melpa/commit/4872ef038dbbf67008bfa7951574ee372d6ff68d.